% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-bigtabulate.R
\name{bigtabulate-package}
\alias{bigtabulate-package}
\title{Extended Tabular Operations for Both matrix and big.matrix Objects}
\description{
The \pkg{bigmemory} package contains the core support and 
basic summary functions; \pkg{bigtabulate} contains extended tabulation 
functionality.
}
\author{
Michael J. Kane and John W. Emerson

Maintainers Michael J. Kane <bigmemoryauthors@gmail.com>
}
\references{
\url{http://www.bigmemory.org}
}
\keyword{package}

