\name{bigtabulate-package}
\alias{bigtabulate-package}
\docType{package}
\title{
A suite of functions for fast, efficient tabular operations (table-,
tapply-, split-, and aggregate-like) with both matrix and big.matrix
objects.
}
\description{
See \code{\link{bigtabulate}}, \code{\link{bigtable}}, \code{\link{bigsplit}},
\code{\link{bigaggregate}}, and \code{\link{bigtsummary}}.
}
\details{
\tabular{ll}{
Package: \tab bigtabulate\cr
Type: \tab Package\cr
Version: \tab 1.0.5\cr
Date: \tab 2010-04-08\cr
License: \tab LGPL-3\cr
Copyright: \tab (C) 2010 John W. Emerson and Michael J. Kane\cr
URL: \tab http://www.bigmemory.org\cr
LazyLoad: \tab yes\cr
}

The \pkg{bigmemory} package contains the core support and basic summary
functions; \pkg{bigtabulate} contains extended tabulation functionality.

}
\author{
John W. Emerson and Michael J. Kane

Maintainers: Jay and Mike <bigmemoryauthors@gmail.com>
}
\keyword{package}
\references{The Bigmemory Project: \url{http://www.bigmemory.org/}.}
\seealso{For example, see
\code{\link[bigmemory]{big.matrix}}, \code{\link[biglm]{biglm}},
\code{\link[biganalytics]{bigkmeans}}, \code{\link[biganalytics]{binit}}, \code{\link[biganalytics]{colmean}}, \code{\link{tapply}}.
}
\examples{
# None provided here.
}
